display preview has it's on file now and  is imported at application.js. Can preview both, image and video, but only one still
