+++ app/javascript/packs/application.js
@@ -35,45 +35,13 @@ import { initChatroomCable } from '../channels/chatroom_channel'
 
 import { initMapbox } from '../plugins/init_mapbox';
 
-
 import { initProfileNav } from '../plugins/profile_nav_item'
 
 import { initAutocomplete } from '../plugins/init_autocomplete';
 
-
-// Code to add preview when uploading file
-const input = document.getElementById('post_photos')
-const previews = document.getElementById('previews')
+import { displayPreview } from '../plugins/displayPreview'
 
 
-// se quiser improve pode display more
-
-const displayPreview = (input) => {
-  if (input.files && input.files[0]) {
-    const reader = new FileReader();
-    reader.onload = (event) => {
-      previews.insertAdjacentHTML('beforeend', `<img width="100" height="100" id="photo-preview" src="${event.currentTarget.result}">`)
-      // document.getElementById('photo-preview').src = event.currentTarget.result;
-    }
-    reader.readAsDataURL(input.files[0])
-    document.getElementById('photo-preview').classList.remove('hidden');
-  }
-}
-
-if(input){
-  input.addEventListener('change', (event) => {
-    if(input.files[0].type.split("/")[0] == 'image'){
-      console.log(input.files[0].type.split("/")[0] == 'image')
-      displayPreview(input)
-    }
-  });
-};
-
-// for (var i = 0; i < files.length; i++) {
-//   var name = files[i].name;
-//   var type = files[i].type;
-//   console.log("Filename: " + name + " , Type: " + type);
-// }
 
 
 document.addEventListener('turbolinks:load', () => {
@@ -89,6 +57,8 @@ document.addEventListener('turbolinks:load', () => {
   initProfileNav();
 
   initAutocomplete();
+
+  displayPreview()
 
 });
